fix common js errors from logs. reorg js into modules
