+++ src/scripts/components/table-of-contents.js
@@ -1,121 +1,46 @@
-$(document).ready(function () {
+import { moveToAnchor } from '../helpers/moveToAnchor';
 
-    function hideToc()
-    {
-        // hide toc
-        $('.toc-container > div').hide();
+const largeScreenThreshold = 1710;
+var sidenavMapping = [];
+var apiNavMapping = [];
 
-        // hide mobile toc button
-        $('.mobile-toc-toggle').removeClass('d-block').addClass('d-none');
-    }
 
-    if($('#TableOfContents ul').length) {
-        // when page ready collect mapping of link to headers so we aren't checking the dom all the time
-        var mapping = [];
+if (!$('#TableOfContents ul').length) {
+    $('.mobile-toc-toggle').addClass('d-none');
+    $('.toc-container').addClass('d-none');
+}
 
-        function buildMap() {
-            mapping = [];
-            var link = null;
-            $('#TableOfContents ul a').each(function() {
-                var href = $(this).attr('href');
-                var id = href.replace('#', '').replace(' ','-');
-                var header = $('[id="'+id+'"]');
-                var navParentLinks = $(this).parents('#TableOfContents').find('ul > li').has($(this)).find('> a');
+// when the page loads, checking the window width
+widthCheck();
+tocWidthUpdate();
 
-                if(header.length) {
-                    if(header.is('h2') || header.is('h3')) {
-                        mapping.push({
-                            'navLink': $(this),
-                            'navLinkPrev': link,
-                            'navParentLinks': navParentLinks,
-                            'id': id,
-                            'header': header,
-                            'isH2': header.is('h2'),
-                            'isH3': header.is('h3')
-                        });
-                    }
-                }
-                link = $(this);
-            });
-        }
+if(innerWidth > largeScreenThreshold && window.scrollY < 60){
+    $('.toc').css('top', 189);
+    $('.mobile-toc-toggle').css('top', 189);
+}else if(innerWidth > largeScreenThreshold && window.scrollY > 60){
+    $('.toc').css('top', 124.5);
+    $('.mobile-toc-toggle').css('top', 124.5);
+}
 
-        function onScroll() {
-            var winTop = $(window).scrollTop();
-            var localOffset = 120;
+$(window).on('resize scroll', function(e) {
+    var header_h = $('body > header').height();
+    var top = $('#TableOfContents').position() ? $('#TableOfContents').position().top : 0;
+    var offset = header_h + top;
+    $('.toc').css('maxHeight', document.documentElement.clientHeight - offset);
+});
 
-            if($(window).scrollTop() + $(window).height() === $(document).height()) {
-                // we are at the bottom of the screen  just highlight the last item we can
-                $('.toc_open').removeClass('toc_open');
-                $('.toc_scrolled').removeClass('toc_scrolled');
-                var obj = mapping[mapping.length-1];
-                if(obj) {
-                    if(obj.isH3) {
-                        obj.navParentLinks.each(function() {
-                            var href = $(this).attr('href');
-                            var id = href.replace('#', '').replace(' ','-');
-                            var header = $('[id="'+id+'"]');
-                            if(header.is('h2')) {
-                                $(this).addClass('toc_open');
-                            }
-                        });
-                        obj.navLink.parent().addClass('toc_scrolled');
-                    }
-                    if(obj.isH2) {
-                        obj.navLink.parent().addClass('toc_scrolled');
-                    }
-                }
-            } else {
-                $('.toc_open').removeClass('toc_open');
-                for(var i = 0; i < mapping.length; i++) {
-                    var obj = mapping[i];
-                    var j = i+1;
-                    if(j > mapping.length) { j = 0; }
-                    var nextobj = mapping[j];
-                    obj.navLink.parent().removeClass('toc_scrolled');
+$(window).on('resize', function() {
+    widthCheck();
+    tocWidthUpdate();
+});
 
-                    if( (winTop >= obj.header.offset().top - localOffset) && (typeof(nextobj) === 'undefined' || winTop < nextobj.header.offset().top - localOffset) ) {
-                        obj.navLink.parent().addClass('toc_scrolled');
-                        // add toc open to parents of this toc_scrolled
-                        obj.navParentLinks.each(function() {
-                            var href = $(this).attr('href');
-                            var id = href.replace('#', '').replace(' ','-');
-                            var header = $('[id="'+id+'"]');
-                            if(header.is('h2')) {
-                                $(this).addClass('toc_open');
-                            }
-                        });
-                    }
-                }
-            }
-        }
-
-        $('.mobile-toc-toggle').on('click touch', function () {
-            var icon = $(this).find('i');
-            var open = icon.hasClass('icon-small-x');
-            if(open) {
-                $('.toc-container').toggleClass('mobile-open').toggleClass('d-none');
-            } else {
-                $('.toc-container').toggleClass('mobile-open').toggleClass('d-none');
-            }
-            $(this).find('i').toggleClass('icon-small-x').toggleClass('icon-small-bookmark');
-            $( document ).trigger( "headerResize", [ parseInt($('body > header').height()) ] );
-        });
-
-        $(document).on( "moveToAnchor", function() {
-            var open = $('.mobile-toc-toggle i').hasClass('icon-small-x');
-            if(open) {
-                //$('.mobile-toc-toggle').click();
-            }
-        });
-
-        // Setup for large screen ToC
-        var largeScreenThreshold = 1710;
-
-        $(window).on('resize scroll', function(e) {
-            onScroll();
-            if($(window).width() > 530 && $(window).width() < largeScreenThreshold) {
-                var bottomOfBrowser = parseInt($(document).scrollTop()) + parseInt($(window).height());
-                var footerTop = $('body > footer').offset().top;
+$(window).on('resize scroll', function(e) {
+    onScroll();
+    if($(window).width() > 530 && $(window).width() < largeScreenThreshold) {
+        var bottomOfBrowser = parseInt($(document).scrollTop()) + parseInt($(window).height());
+        var footerTop = $('body > footer').offset().top;
+        if (!$('.mobile-toc-toggle').hasClass('js-disable-toc')) {
+            if ($('#TableOfContents ul').length) {
                 if(bottomOfBrowser >= footerTop) {
                     if(!$('.mobile-toc-toggle').hasClass('d-none')) {
                         $('.mobile-toc-toggle').toggleClass('d-none');
@@ -129,27 +54,34 @@ $(document).ready(function () {
                     }
                 }
             }
-        }).trigger('scroll');
-
-        $(document).on( "headerResize", function( event, height) {
-            var offset = 30;
-            if($('.announcement_banner.open').length) {
-                offset = 60;
-            }
-            $('.mobile-toc-toggle').css('top', height + offset + 'px');
-            $('.toc').css('top', height + offset + 'px');
-        });
-
-        buildMap();
-        onScroll();
-    } else {
-        hideToc();
+        } else {
+            $('.mobile-toc-toggle').addClass('d-none');
+            $('.toc-container').addClass('d-none');
+        }
     }
+}).trigger('scroll');
 
-    // hiding + displaying the ToC depending on the window width
-    function widthCheck(){
-        // if ToC elements exist
-        if($('#TableOfContents ul').length) {
+$(document).on( "headerResize", function( event, height) {
+    var offset = 30;
+    if($('.announcement_banner.open').length) {
+        offset = 60;
+    }
+    $('.mobile-toc-toggle').css('top', height + offset + 'px');
+    $('.toc').css('top', height + offset + 'px');
+});
+
+export function hideToc(){
+    // hide toc
+    $('.toc-container > div').hide();
+
+    // hide mobile toc button
+    $('.mobile-toc-toggle').removeClass('d-block').addClass('d-none');
+}
+
+// hiding + displaying the ToC depending on the window width
+export function widthCheck(){
+    if (!$('.toc-container').hasClass('js-disable-toc')) {
+        if ($('#TableOfContents ul').length) {
             if(innerWidth > largeScreenThreshold){
                 if($('.toc-container').hasClass('d-none') && $('.mobile-toc-toggle i').hasClass('icon-small-bookmark')){
                     $('.toc-container').toggleClass('mobile-open').toggleClass('d-none');
@@ -162,36 +94,148 @@ $(document).ready(function () {
                 }
             }
         }
+
+    } else {
+        $('.mobile-toc-toggle').addClass('d-none');
     }
 
-    // updating ToC width on large screens
-    function tocWidthUpdate(){
-        if(innerWidth > largeScreenThreshold){
-            $('.toc-container.mobile-open').css("width", (innerWidth/2)-600 + "px");
+}
+
+// updating ToC width on large screens
+export function tocWidthUpdate(){
+    if(innerWidth > largeScreenThreshold){
+        $('.toc-container.mobile-open').css("width", (innerWidth/2)-600 + "px");
+    }
+}
+
+// -------------- end TOC ---------------
+
+export function showTOCIcon(){
+    // $('.toc-container').addClass('mobile-open').removeClass('d-none');
+    $('.mobile-toc-toggle').removeClass('d-none');
+    $(this).find('i').addClass('icon-small-x').removeClass('icon-small-bookmark');
+    $( document ).trigger( "headerResize", [ parseInt($('body > header').height()) ] );
+}
+
+export function updateTOC(){
+
+    $('.toc').css('display', 'block');
+    $('#TableOfContents a').on('click', function(e) {
+        var href = $(this).attr('href');
+        if(href.substr(0, 1) === '#') {
+            moveToAnchor(href.substr(1), true);
+            return false;
         }
-    }
-
-    // when the page loads, checking the window width
-    widthCheck();
-    tocWidthUpdate();
-    
-    if(innerWidth > largeScreenThreshold && window.scrollY < 60){
-        $('.toc').css('top', '189px');
-        $('.mobile-toc-toggle').css('top', '189px');
-    }else if(innerWidth > largeScreenThreshold && window.scrollY > 60){
-        $('.toc').css('top', '124.5px');
-        $('.mobile-toc-toggle').css('top', '124.5px');
-    }
-
-    $(window).on('resize scroll', function(e) {
-        var header_h = $('body > header').height();
-        var top = $('#TableOfContents').position() ? $('#TableOfContents').position().top : 0;
-        var offset = header_h + top;
-        $('.toc').css('maxHeight', document.documentElement.clientHeight - offset);
     });
 
-    $(window).on('resize', function() {
-        widthCheck();
-        tocWidthUpdate();
+}
+
+export function buildTOCMap() {
+    sidenavMapping = [];
+    var link = null;
+    $('#TableOfContents ul a').each(function() {
+        var href = $(this).attr('href');
+        var id = href.replace('#', '').replace(' ','-');
+        var header = $('[id="'+id+'"]');
+        var navParentLinks = $(this).parents('#TableOfContents').find('ul > li').has($(this)).find('> a');
+
+        if(header.length) {
+            if(header.is('h2') || header.is('h3')) {
+                sidenavMapping.push({
+                    'navLink': $(this),
+                    'navLinkPrev': link,
+                    'navParentLinks': navParentLinks,
+                    'id': id,
+                    'header': header,
+                    'isH2': header.is('h2'),
+                    'isH3': header.is('h3')
+                });
+            }
+        }
+        link = $(this);
     });
-});
+}
+
+export function buildAPIMap() {
+    apiNavMapping = [];
+    var link = null;
+    $('.sidenav-api ul a').each(function() {
+        var href = $(this).attr('href');
+        var id = href.replace('#', '').replace(' ','-');
+        var header = $('[id="'+id+'"]');
+        var navParentLinks = $(this).parents('.sidenav-api').find('ul > li').has($(this)).find('> a');
+
+        if(header.length) {
+            if(header.is('h2') || header.is('h3')) {
+                apiNavMapping.push({
+                    'navLink': $(this),
+                    'navLinkPrev': link,
+                    'navParentLinks': navParentLinks,
+                    'id': id,
+                    'header': header,
+                    'isH2': header.is('h2'),
+                    'isH3': header.is('h3')
+                });
+            }
+        } else {
+            //console.log('could not find h2[id="'+id+'"]');
+        }
+        link = $(this);
+    });
+}
+
+export function onScroll() {
+    var winTop = $(window).scrollTop();
+    var localOffset = 120;
+
+    if($(window).scrollTop() + $(window).height() === $(document).height()) {
+        // we are at the bottom of the screen  just highlight the last item
+    } else {
+        $('.toc_open').removeClass('toc_open');
+
+        // tocMapping
+        for(var i = 0; i < sidenavMapping.length; i++) {
+            var obj = sidenavMapping[i];
+            var j = i+1;
+            if(j > sidenavMapping.length) { j = 0; }
+            var nextobj = sidenavMapping[j];
+            obj.navLink.removeClass('toc_scrolled');
+
+            if( (winTop >= obj.header.offset().top - localOffset) && (typeof(nextobj) === 'undefined' || winTop < nextobj.header.offset().top - localOffset) ) {
+                obj.navLink.addClass('toc_scrolled');
+                // add toc open to parents of this toc_scrolled
+                obj.navParentLinks.each(function() {
+                    var href = $(this).attr('href');
+                    var id = href.replace('#', '').replace(' ','-');
+                    var header = $('[id="'+id+'"]');
+                    if(header.is('h2')) {
+                        $(this).addClass('toc_open');
+                    }
+                });
+            }
+        }
+
+        // apiMapping
+        for(var i = 0; i < apiNavMapping.length; i++) {
+            var obj = apiNavMapping[i];
+            var j = i+1;
+            if(j > apiNavMapping.length) { j = 0; }
+            var nextobj = apiNavMapping[j];
+            obj.navLink.removeClass('toc_scrolled');
+
+            if( (winTop >= obj.header.offset().top - localOffset) && (typeof(nextobj) === 'undefined' || winTop < nextobj.header.offset().top - localOffset) ) {
+                obj.navLink.addClass('toc_scrolled');
+                // add toc open to parents of this toc_scrolled
+                obj.navParentLinks.each(function() {
+                    var href = $(this).attr('href');
+                    var id = href.replace('#', '').replace(' ','-');
+                    var header = $('[id="'+id+'"]');
+                    if(header.is('h2')) {
+                        $(this).addClass('toc_open');
+                    }
+                });
+            }
+        }
+
+    }
+}
