+++ src/scripts/main-dd-js.js
@@ -9,6 +9,4 @@ import './components/header';
 import './components/platforms';
 
 // TODO: split up code from datadog-docs.js into modules after webpack migration
-// import './components/integrations';
 // import './components/sidenav';
-// import './components/table-of-contents';
